Move zebra_state::service::check tests to their own module
