+++ zebra-state/src/service/check/tests.rs
@@ -1,0 +1,3 @@
+//! Tests for state contextual validation checks.
+
+mod vectors;
